+++ tests/test_gateway_mqtt.py
@@ -1,210 +1,234 @@
 """Test mysensors MQTT gateway with unittest."""
-import os
-import tempfile
+import logging
 import time
-from unittest import TestCase, main, mock
+from unittest import mock
+
+import pytest
 
 from mysensors.gateway_mqtt import MQTTGateway
 from mysensors.persistence import Persistence
-from mysensors.sensor import ChildSensor, Sensor
+from mysensors.sensor import Sensor
+
+# pylint: disable=redefined-outer-name, too-many-arguments
 
 
-class TestMQTTGateway(TestCase):
-    """Test the MQTT Gateway."""
+@pytest.fixture
+def mock_pub():
+    """Return a mock callback to publish to mqtt broker."""
+    return mock.Mock()
 
-    def setUp(self):
-        """Set up gateway."""
-        self.mock_pub = mock.Mock()
-        self.mock_sub = mock.Mock()
-        self.gateway = MQTTGateway(self.mock_pub, self.mock_sub)
 
-    def tearDown(self):
-        """Stop MQTTGateway if alive."""
-        self.gateway.stop()
+@pytest.fixture
+def mock_sub():
+    """Return a mock callback to subscribe to a mqtt topic."""
+    return mock.Mock()
 
-    def _add_sensor(self, sensorid):
+
+@pytest.fixture
+def gateway(mock_pub, mock_sub):
+    """Yield gateway instance."""
+    _gateway = MQTTGateway(mock_pub, mock_sub)
+    yield _gateway
+    _gateway.stop()
+
+
+def get_gateway(*args, **kwargs):
+    """Return a gateway instance."""
+    return MQTTGateway(*args, **kwargs)
+
+
+@pytest.fixture
+def add_sensor(gateway):
+    """Return function for adding node."""
+    def _add_sensor(sensor_id):
         """Add sensor node. Return sensor node instance."""
-        self.gateway.sensors[sensorid] = Sensor(sensorid)
-        return self.gateway.sensors[sensorid]
-
-    def test_send(self):
-        """Test send method."""
-        self.gateway.send('1;1;1;0;1;20\n')
-        self.mock_pub.assert_called_with('/1/1/1/0/1', '20', 0, True)
-
-    def test_send_empty_string(self):
-        """Test send method with empty string."""
-        self.gateway.send('')
-        self.assertFalse(self.mock_pub.called)
-
-    def test_send_error(self):
-        """Test send method with error on publish."""
-        self.mock_pub.side_effect = ValueError(
-            'Publish topic cannot contain wildcards.')
-        with self.assertLogs(level='ERROR') as test_handle:
-            self.gateway.send('1;1;1;0;1;20\n')
-        self.mock_pub.assert_called_with('/1/1/1/0/1', '20', 0, True)
-        self.assertEqual(
-            # only check first line of error log
-            test_handle.output[0].split('\n', 1)[0],
-            'ERROR:mysensors.gateway_mqtt:Publish to /1/1/1/0/1 failed: '
-            'Publish topic cannot contain wildcards.')
-
-    def test_recv(self):
-        """Test recv method."""
-        sensor = self._add_sensor(1)
-        sensor.children[1] = ChildSensor(
-            1, self.gateway.const.Presentation.S_HUM)
-        sensor.children[1].values[self.gateway.const.SetReq.V_HUM] = '20'
-        self.gateway.recv('/1/1/2/0/1', '', 0)
-        ret = self.gateway.run_job()
-        self.assertEqual(ret, '1;1;1;0;1;20\n')
-        self.gateway.recv('/1/1/2/0/1', '', 1)
-        ret = self.gateway.run_job()
-        self.assertEqual(ret, '1;1;1;1;1;20\n')
-
-    def test_recv_wrong_prefix(self):
-        """Test recv method with wrong topic prefix."""
-        sensor = self._add_sensor(1)
-        sensor.children[1] = ChildSensor(
-            1, self.gateway.const.Presentation.S_HUM)
-        sensor.children[1].values[self.gateway.const.SetReq.V_HUM] = '20'
-        self.gateway.recv('wrong/1/1/2/0/1', '', 0)
-        ret = self.gateway.run_job()
-        self.assertEqual(ret, None)
-
-    def test_presentation(self):
-        """Test handle presentation message."""
-        self._add_sensor(1)
-        self.gateway.logic('1;1;0;0;7;Humidity Sensor\n')
-        calls = [
-            mock.call('/1/1/1/+/+', self.gateway.recv, 0),
-            mock.call('/1/1/2/+/+', self.gateway.recv, 0),
-            mock.call('/1/+/4/+/+', self.gateway.recv, 0)]
-        self.mock_sub.assert_has_calls(calls)
-
-    def test_presentation_no_sensor(self):
-        """Test handle presentation message without sensor."""
-        self.gateway.logic('1;1;0;0;7;Humidity Sensor\n')
-        self.assertFalse(self.mock_sub.called)
-
-    def test_subscribe_error(self):
-        """Test subscribe throws error."""
-        self._add_sensor(1)
-        self.mock_sub.side_effect = ValueError(
-            'No topic specified, or incorrect topic type.')
-        with self.assertLogs(level='ERROR') as test_handle:
-            self.gateway.logic('1;1;0;0;7;Humidity Sensor\n')
-        calls = [
-            mock.call('/1/1/1/+/+', self.gateway.recv, 0),
-            mock.call('/1/1/2/+/+', self.gateway.recv, 0)]
-        self.mock_sub.assert_has_calls(calls)
-        self.assertEqual(
-            # only check first line of error log
-            test_handle.output[0].split('\n', 1)[0],
-            'ERROR:mysensors.gateway_mqtt:Subscribe to /1/1/1/+/+ failed: '
-            'No topic specified, or incorrect topic type.')
-
-    @mock.patch('mysensors.persistence.Persistence.safe_load_sensors')
-    @mock.patch('mysensors.persistence.Persistence.save_sensors')
-    def test_start_stop_gateway(self, mock_save, mock_load):
-        """Test start and stop of MQTT gateway."""
-        self.gateway.persistence = Persistence(self.gateway.sensors)
-        mock_cancel_save = mock.MagicMock()
-        mock_schedule_save = mock.MagicMock()
-        mock_schedule_save.return_value = mock_cancel_save
-        self.gateway.persistence.schedule_save_sensors = mock_schedule_save
-        sensor = self._add_sensor(1)
-        sensor.children[1] = ChildSensor(
-            1, self.gateway.const.Presentation.S_HUM)
-        sensor.children[1].values[self.gateway.const.SetReq.V_HUM] = '20'
-        self.gateway.recv('/1/1/2/0/1', '', 0)
-        self.gateway.recv('/1/1/1/0/1', '30', 0)
-        self.gateway.recv('/1/1/2/0/1', '', 0)
-        self.gateway.start_persistence()
-        assert mock_load.call_count == 1
-        assert mock_schedule_save.call_count == 1
-        self.gateway.start()
-        time.sleep(0.05)
-        calls = [
-            mock.call('/+/+/0/+/+', self.gateway.recv, 0),
-            mock.call('/+/+/3/+/+', self.gateway.recv, 0)]
-        self.mock_sub.assert_has_calls(calls)
-        calls = [
-            mock.call('/1/1/1/0/1', '20', 0, True),
-            mock.call('/1/1/1/0/1', '30', 0, True)]
-        self.mock_pub.assert_has_calls(calls)
-        self.gateway.stop()
-        assert mock_cancel_save.call_count == 1
-        assert mock_save.call_count == 1
-
-    def test_mqtt_load_persistence(self):
-        """Test load persistence file for MQTTGateway."""
-        sensor = self._add_sensor(1)
-        sensor.children[1] = ChildSensor(
-            1, self.gateway.const.Presentation.S_HUM)
-        sensor.children[1].values[self.gateway.const.SetReq.V_HUM] = '20'
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            persistence_file = os.path.join(temp_dir, 'file.json')
-            self.gateway.persistence = Persistence(
-                self.gateway.sensors, persistence_file)
-            self.gateway.persistence.save_sensors()
-            del self.gateway.sensors[1]
-            self.assertNotIn(1, self.gateway.sensors)
-            self.gateway.persistence.safe_load_sensors()
-            # pylint: disable=protected-access
-            self.gateway._init_topics()
-        self.assertEqual(
-            self.gateway.sensors[1].children[1].id,
-            sensor.children[1].id)
-        self.assertEqual(
-            self.gateway.sensors[1].children[1].type,
-            sensor.children[1].type)
-        self.assertEqual(
-            self.gateway.sensors[1].children[1].values,
-            sensor.children[1].values)
-        calls = [
-            mock.call('/1/1/1/+/+', self.gateway.recv, 0),
-            mock.call('/1/1/2/+/+', self.gateway.recv, 0),
-            mock.call('/1/+/4/+/+', self.gateway.recv, 0)]
-        self.mock_sub.assert_has_calls(calls)
+        gateway.sensors[sensor_id] = Sensor(sensor_id)
+        return gateway.sensors[sensor_id]
+    return _add_sensor
 
 
-class TestMQTTGatewayCustomPrefix(TestCase):
-    """Test the MQTT Gateway with custom topic prefix."""
-
-    def setUp(self):
-        """Set up test."""
-        self.mock_pub = mock.Mock()
-        self.mock_sub = mock.Mock()
-        self.gateway = None
-
-    def _setup(self, in_prefix, out_prefix):
-        """Set up gateway."""
-        self.gateway = MQTTGateway(
-            self.mock_pub, self.mock_sub, in_prefix=in_prefix,
-            out_prefix=out_prefix)
-
-    def _add_sensor(self, sensorid):
-        """Add sensor node. Return sensor node instance."""
-        self.gateway.sensors[sensorid] = Sensor(sensorid)
-        return self.gateway.sensors[sensorid]
-
-    def test_nested_prefix(self):
-        """Test recv method with nested topic prefix."""
-        self._setup('test/test-in', 'test/test-out')
-        sensor = self._add_sensor(1)
-        sensor.children[1] = ChildSensor(
-            1, self.gateway.const.Presentation.S_HUM)
-        sensor.children[1].values[self.gateway.const.SetReq.V_HUM] = '20'
-        self.gateway.recv('test/test-in/1/1/2/0/1', '', 0)
-        ret = self.gateway.run_job()
-        self.assertEqual(ret, '1;1;1;0;1;20\n')
-        self.gateway.recv('test/test-in/1/1/2/0/1', '', 1)
-        ret = self.gateway.run_job()
-        self.assertEqual(ret, '1;1;1;1;1;20\n')
+def get_sensor(sensor_id, gateway):
+    """Add sensor on gateway and return sensor instance."""
+    gateway.sensors[sensor_id] = Sensor(sensor_id)
+    return gateway.sensors[sensor_id]
 
 
-if __name__ == '__main__':
-    main()
+def test_send(gateway, mock_pub):
+    """Test send method."""
+    gateway.send('1;1;1;0;1;20\n')
+    assert mock_pub.call_count == 1
+    assert mock_pub.call_args == mock.call('/1/1/1/0/1', '20', 0, True)
+
+
+def test_send_empty_string(gateway, mock_pub):
+    """Test send method with empty string."""
+    gateway.send('')
+    assert mock_pub.call_count == 0
+
+
+def test_send_error(gateway, mock_pub, caplog):
+    """Test send method with error on publish."""
+    mock_pub.side_effect = ValueError(
+        'Publish topic cannot contain wildcards.')
+    caplog.set_level(logging.ERROR)
+    gateway.send('1;1;1;0;1;20\n')
+    assert mock_pub.call_count == 1
+    assert mock_pub.call_args == mock.call('/1/1/1/0/1', '20', 0, True)
+    assert (
+        'Publish to /1/1/1/0/1 failed: '
+        'Publish topic cannot contain wildcards.' in caplog.text)
+
+
+def test_recv(gateway, add_sensor):
+    """Test recv method."""
+    sensor = add_sensor(1)
+    sensor.add_child_sensor(1, gateway.const.Presentation.S_HUM)
+    sensor.children[1].values[gateway.const.SetReq.V_HUM] = '20'
+    gateway.recv('/1/1/2/0/1', '', 0)
+    ret = gateway.run_job()
+    assert ret == '1;1;1;0;1;20\n'
+    gateway.recv('/1/1/2/0/1', '', 1)
+    ret = gateway.run_job()
+    assert ret == '1;1;1;1;1;20\n'
+
+
+def test_recv_wrong_prefix(gateway, add_sensor):
+    """Test recv method with wrong topic prefix."""
+    sensor = add_sensor(1)
+    sensor.add_child_sensor(1, gateway.const.Presentation.S_HUM)
+    sensor.children[1].values[gateway.const.SetReq.V_HUM] = '20'
+    gateway.recv('wrong/1/1/2/0/1', '', 0)
+    ret = gateway.run_job()
+    assert ret is None
+
+
+def test_presentation(gateway, add_sensor, mock_sub):
+    """Test handle presentation message."""
+    add_sensor(1)
+    gateway.logic('1;1;0;0;7;Humidity Sensor\n')
+    calls = [
+        mock.call('/1/1/1/+/+', gateway.recv, 0),
+        mock.call('/1/1/2/+/+', gateway.recv, 0),
+        mock.call('/1/+/4/+/+', gateway.recv, 0)]
+    assert mock_sub.call_count == 3
+    assert mock_sub.mock_calls == calls
+
+
+def test_presentation_no_sensor(gateway, mock_sub):
+    """Test handle presentation message without sensor."""
+    gateway.logic('1;1;0;0;7;Humidity Sensor\n')
+    assert mock_sub.call_count == 0
+
+
+def test_subscribe_error(gateway, add_sensor, mock_sub, caplog):
+    """Test subscribe throws error."""
+    add_sensor(1)
+    mock_sub.side_effect = ValueError(
+        'No topic specified, or incorrect topic type.')
+    caplog.set_level(logging.ERROR)
+    gateway.logic('1;1;0;0;7;Humidity Sensor\n')
+    calls = [
+        mock.call('/1/1/1/+/+', gateway.recv, 0),
+        mock.call('/1/1/2/+/+', gateway.recv, 0),
+        mock.call('/1/+/4/+/+', gateway.recv, 0)]
+    assert mock_sub.call_count == 3
+    assert mock_sub.mock_calls == calls
+    assert (
+        'Subscribe to /1/1/1/+/+ failed: '
+        'No topic specified, or incorrect topic type.' in caplog.text)
+
+
+@mock.patch('mysensors.persistence.Persistence.safe_load_sensors')
+@mock.patch('mysensors.persistence.Persistence.save_sensors')
+def test_start_stop_gateway(
+        mock_save, mock_load, gateway, add_sensor, mock_pub, mock_sub):
+    """Test start and stop of MQTT gateway."""
+    mock_schedule_factory = mock.MagicMock()
+    mock_schedule_save = mock.MagicMock()
+    mock_schedule_factory.return_value = mock_schedule_save
+    gateway.persistence = Persistence(gateway.sensors, mock_schedule_factory)
+    sensor = add_sensor(1)
+    sensor.add_child_sensor(1, gateway.const.Presentation.S_HUM)
+    sensor.children[1].values[gateway.const.SetReq.V_HUM] = '20'
+    gateway.recv('/1/1/2/0/1', '', 0)  # should generate a publish of 20
+    gateway.recv('/1/1/1/0/1', '30', 0)
+    gateway.recv('/1/1/2/0/1', '', 0)  # should generate a publish of 30
+    gateway.start_persistence()
+    assert mock_load.call_count == 1
+    assert mock_schedule_save.call_count == 1
+    gateway.start()
+    time.sleep(0.05)
+    calls = [
+        mock.call('/+/+/0/+/+', gateway.recv, 0),
+        mock.call('/+/+/3/+/+', gateway.recv, 0),
+        mock.call('/1/1/1/+/+', gateway.recv, 0),
+        mock.call('/1/1/2/+/+', gateway.recv, 0),
+        mock.call('/1/+/4/+/+', gateway.recv, 0)]
+    assert mock_sub.call_count == 5
+    assert mock_sub.mock_calls == calls
+    calls = [
+        mock.call('/1/1/1/0/1', '20', 0, True),
+        mock.call('/1/1/1/0/1', '30', 0, True)]
+    assert mock_pub.call_count == 2
+    assert mock_pub.mock_calls == calls
+    gateway.stop()
+    assert mock_save.call_count == 1
+
+
+def test_mqtt_load_persistence(gateway, add_sensor, mock_sub, tmpdir):
+    """Test load persistence file for MQTTGateway."""
+    sensor = add_sensor(1)
+    sensor.add_child_sensor(1, gateway.const.Presentation.S_HUM)
+    sensor.children[1].values[gateway.const.SetReq.V_HUM] = '20'
+
+    persistence_file = tmpdir.join('file.json')
+    gateway.persistence = Persistence(
+        gateway.sensors, mock.MagicMock(), persistence_file.strpath)
+    gateway.persistence.save_sensors()
+    del gateway.sensors[1]
+    assert 1 not in gateway.sensors
+    gateway.persistence.safe_load_sensors()
+    # pylint: disable=protected-access
+    gateway._init_topics()
+    assert gateway.sensors[1].children[1].id == sensor.children[1].id
+    assert gateway.sensors[1].children[1].type == sensor.children[1].type
+    assert gateway.sensors[1].children[1].values == sensor.children[1].values
+    calls = [
+        mock.call('/+/+/0/+/+', gateway.recv, 0),
+        mock.call('/+/+/3/+/+', gateway.recv, 0),
+        mock.call('/1/1/1/+/+', gateway.recv, 0),
+        mock.call('/1/1/2/+/+', gateway.recv, 0),
+        mock.call('/1/+/4/+/+', gateway.recv, 0)]
+    assert mock_sub.call_count == 5
+    assert mock_sub.mock_calls == calls
+
+
+def test_nested_prefix(mock_pub, mock_sub):
+    """Test recv and send method with nested topic prefix."""
+    gateway = get_gateway(
+        mock_pub, mock_sub, in_prefix='test/test-in',
+        out_prefix='test/test-out')
+    sensor = get_sensor(1, gateway)
+    sensor.add_child_sensor(1, gateway.const.Presentation.S_HUM)
+    sensor.children[1].values[gateway.const.SetReq.V_HUM] = '20'
+    gateway.recv('test/test-in/1/1/2/0/1', '', 0)
+    ret = gateway.run_job()
+    assert ret == '1;1;1;0;1;20\n'
+    gateway.send(ret)
+    assert mock_pub.call_args == mock.call(
+        'test/test-out/1/1/1/0/1', '20', 0, True)
+    gateway.recv('test/test-in/1/1/2/0/1', '', 1)
+    ret = gateway.run_job()
+    assert ret == '1;1;1;1;1;20\n'
+    gateway.send(ret)
+    assert mock_pub.call_args == mock.call(
+        'test/test-out/1/1/1/1/1', '20', 1, True)
+
+
+def test_get_gateway_id(mock_pub, mock_sub):
+    """Test get_gateway_id method."""
+    gateway = get_gateway(
+        mock_pub, mock_sub, in_prefix='test/test-in',
+        out_prefix='test/test-out')
+    gateway_id = gateway.get_gateway_id()
+    assert gateway_id == 'test/test-in'